tests: add test IDs for mrgvalprep
mrgvalidate/mrgvalprep now expect that tests be tagged with a test ID.
While mrgvalprep::validate_tests() can be told to auto-generate test
IDs if needed, that's just meant to as a backward compatibility bridge
for existing tests don't have IDs.

Aside from the ID naming choices, the main decision that I made here
that should be revisited is the grouping of similar tests under the
same ID.  mrgvalprep allows similar testthat() statements to share the
same ID _if_ they share a common prefix.  So, if I saw a test that
looked to be related variants and already had a shared prefix [*], I
grouped them.  So, it's worth reviewing if 1) something I grouped
would be better kept separate and 2) some other descriptions should be
rewritten so that they can be grouped under the same ID.

[*] For the "writer function - ..." ones, I changed " -" to ":" so
    that "writer function" will show up as the test name in the
    validation docs rather than "writer function -".  mrgvalprep
    automatically strips a trailing ":".